+++ addons/es/tbai/invoices_test.go
@@ -13,7 +13,44 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func validTicketBAIInvoice() *bill.Invoice {
+func TestInvoiceValidation(t *testing.T) {
+	t.Run("standard invoice", func(t *testing.T) {
+		inv := testInvoiceStandard(t)
+		require.NoError(t, inv.Calculate())
+		require.NoError(t, inv.Validate())
+	})
+
+	t.Run("missing customer tax ID", func(t *testing.T) {
+		inv := testInvoiceStandard(t)
+		inv.Customer.TaxID = nil
+		require.NoError(t, inv.Calculate())
+		err := inv.Validate()
+		assert.ErrorContains(t, err, "customer: (tax_id: cannot be blank.)")
+	})
+}
+
+func TestBasqueLineValidation(t *testing.T) {
+	inv := testInvoiceStandard(t)
+	require.NoError(t, inv.Calculate())
+	require.NoError(t, inv.Validate())
+
+	inv.Lines[0].Taxes[0].Ext[tbai.ExtKeyProduct] = "services"
+	require.NoError(t, inv.Calculate())
+	require.NoError(t, inv.Validate())
+
+	inv.Lines[0].Taxes[0].Ext = nil
+	assertValidationError(t, inv, "es-tbai-exemption: required")
+}
+
+func assertValidationError(t *testing.T, inv *bill.Invoice, expected string) {
+	require.NoError(t, inv.Calculate())
+	err := inv.Validate()
+	require.Error(t, err)
+	assert.Contains(t, err.Error(), expected)
+}
+
+func testInvoiceStandard(t *testing.T) *bill.Invoice {
+	t.Helper()
 	return &bill.Invoice{
 		Addons: tax.WithAddons(tbai.V1),
 		Code:   "123",
@@ -51,24 +88,4 @@ func validTicketBAIInvoice() *bill.Invoice {
 			},
 		},
 	}
-}
-
-func TestBasqueLineValidation(t *testing.T) {
-	inv := validTicketBAIInvoice()
-	require.NoError(t, inv.Calculate())
-	require.NoError(t, inv.Validate())
-
-	inv.Lines[0].Taxes[0].Ext[tbai.ExtKeyProduct] = "services"
-	require.NoError(t, inv.Calculate())
-	require.NoError(t, inv.Validate())
-
-	inv.Lines[0].Taxes[0].Ext = nil
-	assertValidationError(t, inv, "es-tbai-exemption: required")
-}
-
-func assertValidationError(t *testing.T, inv *bill.Invoice, expected string) {
-	require.NoError(t, inv.Calculate())
-	err := inv.Validate()
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), expected)
 }