+++ regimes/es/invoices.go
@@ -22,12 +22,7 @@ func (v *invoiceValidator) validate() error {
 	inv := v.inv
 	return validation.ValidateStruct(inv,
 		validation.Field(&inv.Supplier,
-			validation.Required,
 			validation.By(v.supplier),
-			validation.Skip,
-		),
-		validation.Field(&inv.Customer,
-			validation.By(v.customer),
 			validation.Skip,
 		),
 	)
@@ -39,29 +34,11 @@ func (v *invoiceValidator) supplier(value interface{}) error {
 		return nil
 	}
 	return validation.ValidateStruct(obj,
+		// In Spain, all invoices must have a tax ID of the issuer,
+		// there are no exceptions to this rule.
 		validation.Field(&obj.TaxID,
 			validation.Required,
 			tax.RequireIdentityCode,
-			validation.Skip,
-		),
-	)
-}
-
-func (v *invoiceValidator) customer(value interface{}) error {
-	obj, _ := value.(*org.Party)
-	if obj == nil {
-		return nil
-	}
-	// Customers must have a tax ID to at least set the country,
-	// and Spanish ones should also have an ID. There are more complex
-	// rules for exports.
-	return validation.ValidateStruct(obj,
-		validation.Field(&obj.TaxID,
-			validation.Required,
-			validation.When(
-				obj.TaxID != nil && obj.TaxID.Country.In("ES"),
-				tax.RequireIdentityCode,
-			),
 			validation.Skip,
 		),
 	)